+++ src/shapes/Object/defaultValues.ts
@@ -1,4 +1,7 @@
 import { TOP, LEFT } from '../../constants';
+import type { TClassProperties } from '../../typedefs';
+import type { InteractiveFabricObject } from './InteractiveObject';
+import type { FabricObject } from './Object';
 
 export const stateProperties = [
   TOP,
@@ -35,7 +38,9 @@ export const cacheProperties = [
   'clipPath',
 ];
 
-export const fabricObjectDefaultValues = {
+export const fabricObjectDefaultValues: Partial<
+  TClassProperties<FabricObject>
+> = {
   // see composeMatrix() to see order of transforms. First defaults listed based on this
   top: 0,
   left: 0,
@@ -79,7 +84,9 @@ export const fabricObjectDefaultValues = {
   dirty: true,
 } as const;
 
-export const interactiveObjectDefaultValues = {
+export const interactiveObjectDefaultValues: Partial<
+  TClassProperties<InteractiveFabricObject>
+> = {
   noScaleCache: true,
   lockMovementX: false,
   lockMovementY: false,